+++ src/net/Job.cpp
@@ -146,6 +146,10 @@ PowVariant Job::powVariant() const
         } else {
             return PowVariant::POW_V0;
         }
+    } else if (m_powVariant == PowVariant::POW_XTL) {
+        if (m_blob[0] > 5) {
+            return PowVariant::POW_XTL_V9;
+        }
     } else {
         return m_powVariant;
     }